Make Adapter & Persister backwards compatible for implementors
The replacement of the static `name` getter for `Adapter` and
`Persister` in #310 broke custom implementations of these classes
because `id` needed to be overridden. To retain backwards compatibility
we remove the requirement for `id` to be overridden. Instead we delegate
to `name` by default. This does not affect the in-tree implementations.